# Conflicts:
#	premake5.lua
#	src/core/ZoneCull.cpp
#	src/core/Zones.cpp
#	src/objects/CutsceneHead.cpp
#	src/render/Clouds.cpp

+++ src/core/Zones.cpp
@@ -752,9 +752,9 @@ CTheZones::LoadAllZones(uint8 *buffer, uint32 size)
 		meant for a different array, but the game doesn't brake
 		if save data stored is -1.
 		*/
-		MapZoneArray[i].child = GetPointerForZoneIndex((int32)MapZoneArray[i].child);
-		MapZoneArray[i].parent = GetPointerForZoneIndex((int32)MapZoneArray[i].parent);
-		MapZoneArray[i].next = GetPointerForZoneIndex((int32)MapZoneArray[i].next);
+		MapZoneArray[i].child = GetPointerForZoneIndex((uintptr)MapZoneArray[i].child);
+		MapZoneArray[i].parent = GetPointerForZoneIndex((uintptr)MapZoneArray[i].parent);
+		MapZoneArray[i].next = GetPointerForZoneIndex((uintptr)MapZoneArray[i].next);
 		assert(MapZoneArray[i].child == nil);
 		assert(MapZoneArray[i].parent == nil);
 		assert(MapZoneArray[i].next == nil);
